bzero and memset now with size_t

+++ includes/libft.h
@@ -6,13 +6,14 @@
 /*   By: averheij <averheij@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/10/28 13:09:52 by averheij       #+#    #+#                */
-/*   Updated: 2019/10/28 14:19:24 by averheij      ########   odam.nl         */
+/*   Updated: 2019/10/28 15:10:49 by averheij      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIBFT_H
 # define LIBFT_H
+# include <string.h>
 
-void	*ft_memset(void *b, int c, int len);
-void	ft_bzero(void *s, int n);
+void	*ft_memset(void *b, int c, size_t len);
+void	ft_bzero(void *s, size_t n);
 #endif
